Addressed comments - Refactoring - removed excel details from general importer - deleted xlsimporter
Fixes #83
