--- src/main/java/uk/org/tombolo/importer/schools/AbstractSchoolsImporter.java
@@ -1,16 +1,0 @@
-package uk.org.tombolo.importer.schools;
-
-import uk.org.tombolo.core.Provider;
-import uk.org.tombolo.importer.AbstractImporter;
-import uk.org.tombolo.importer.Importer;
-
-/**
- * Abstract class for the school importer indicating the provider
- */
-public abstract class AbstractSchoolsImporter extends XLSImporter {
-    private static final Provider PROVIDER = new Provider("uk.gov.education", "Department of Education");
-
-    public Provider getProvider() {
-        return PROVIDER;
-    }
-}
